+++ src/main/java/com/gmail/nossr50/commands/experience/AddlevelsCommand.java
@@ -22,7 +22,7 @@ public class AddlevelsCommand extends ExperienceCommand {
 
     @Override
     protected void handleCommand(Player player, PlayerProfile profile, PrimarySkillType skill, int value) {
-        float xpRemoved = profile.getSkillXpLevelRaw(skill);
+        double xpRemoved = profile.getSkillXpLevelRaw(skill);
         profile.addLevels(skill, value);
 
         if (player == null) {
